Java and react logic to show edit form
I wanted to implement some logic to show the editForm component when the button is clicked. I wanted the form to be next to that specific student based off of the id from the student. So that way when clicked, the form will display side by side and you can see real time changes.